update CardEditActivity - error, card_edit_text_dialog_title_text settings

+++ app/src/main/java/kr/co/bit/osf/flashcard/CardEditActivity.java
@@ -8,6 +8,7 @@ import android.os.Bundle;
 import android.provider.MediaStore;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
+import android.view.Gravity;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageView;
@@ -36,7 +37,7 @@ public class CardEditActivity extends AppCompatActivity {
     private File photoFile = null;
     private String photoFilePath = null;
     // delete card
-    List<CardDTO>[] cardList = null;
+    List<CardDTO> cardList = null;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -54,26 +55,46 @@ public class CardEditActivity extends AppCompatActivity {
         intentRequestCode = intent.getIntExtra(IntentExtrasName.REQUEST_CODE, 0);
         Dlog.i("intentRequestCode:" + intentRequestCode);
         switch (intentRequestCode) {
+            case IntentRequestCode.CARD_ADD:
+                Dlog.i("intentRequestCode in case:" + intentRequestCode);
+                card = intent.getParcelableExtra(IntentExtrasName.SEND_DATA); //받은 cardDTO에 boxId가 들어있음
+                card.setType(FlashCardDB.CardEntry.TYPE_DEMO);
+                card.setImagePath("" + R.drawable.alphabet_a);
+                card.setName("카드 이름을 입력해주세요");
             case IntentRequestCode.CARD_EDIT:
             case IntentRequestCode.CARD_DELETE:
                 Dlog.i("intentRequestCode in case:" + intentRequestCode);
                 card = intent.getParcelableExtra(IntentExtrasName.SEND_DATA);
+                if(card == null){
+                    Dlog.i("getParcelableExtra: sendData:no data");
+                    finish();
+                    return;
+                }
                 break;
             //delete cardList
-           /* case IntentRequestCode.CARD_DELETE_LIST:
+           case IntentRequestCode.CARD_DELETE_LIST:
                 Dlog.i("intentRequestCode in case:" + intentRequestCode);
-                cardList = intent.getParcelableExtra(IntentExtrasName.SEND_DATA);
-                break;*/
-        }
-        if (card == null && cardList ==null) {
-            Dlog.i("getExtras:sendData:no data");
-            finish();
-            return;
+                cardList = intent.getParcelableArrayListExtra(IntentExtrasName.SEND_DATA);
+                if(cardList == null){
+                    Dlog.i("getParcelableArrayListExtra:sendData: no data");
+                    finish();
+                    return;
+                }
+                break;
         }
 
         // todo: process requested task
         Dlog.i("getExtras:sendData:" + card);
         intentResultCode = RESULT_OK;
+
+        if(intentRequestCode == IntentRequestCode.CARD_ADD){
+            FlashCardDB db = new FlashCardDB(CardEditActivity.this);
+            if(db.addCard(card) == false){
+                intentResultCode = RESULT_CANCELED;
+                Dlog.i("add error:" + card);
+            }
+            finish();
+        }
 
         if(intentRequestCode == IntentRequestCode.CARD_DELETE){
             Dlog.i("delete data:" + card);
@@ -86,30 +107,29 @@ public class CardEditActivity extends AppCompatActivity {
             finish();
         }
         //delete cardList
-/*
         if(intentRequestCode == IntentRequestCode.CARD_DELETE_LIST){
             Dlog.i("delete data:" + cardList);
             //delete cardList
-            FlashCardDB db = new FlashCardDB(CardEditActivity.this);
-            if(db.deleteCard(cardList)==false){
-                intentRequestCode(.getId() == false){
+            if(cardList.size()>0){
+                FlashCardDB db = new FlashCardDB(CardEditActivity.this);
+                if(db.deleteCard(cardList)==false){
                     intentResultCode = RESULT_CANCELED;
                     Dlog.i("delete error:" + card);
+                    finish();
                 }
             }
         }
-*/
 
         imageView = (ImageView) findViewById(R.id.cardEditImageView);
         cardEditTextView = (TextView) findViewById(R.id.cardEditTextView);
 
         //show card
-        if(intentRequestCode == IntentRequestCode.CARD_EDIT) {
+        if(intentRequestCode == IntentRequestCode.CARD_ADD || intentRequestCode == IntentRequestCode.CARD_EDIT) {
             ImageUtil.loadCardImageIntoImageView(this, card, imageView);
         }
         cardEditTextView.setText(card.getName());
 
-        //image view
+        //imageView - card Image
         (findViewById(R.id.cardEditImageView)).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -117,6 +137,7 @@ public class CardEditActivity extends AppCompatActivity {
             }
         });
 
+        //textView - card Name
         (findViewById(R.id.cardEditTextView)).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -198,13 +219,15 @@ public class CardEditActivity extends AppCompatActivity {
 
             AlertDialog.Builder alert = new AlertDialog.Builder(this);
 
-            alert.setTitle(R.string.card_edit_text_dialog_tilte_text);
+            alert.setTitle(R.string.card_edit_text_dialog_title_text);
             alert.setMessage(R.string.card_edit_text_dialog_message_text);
 
             // Set an EditText view to get user input
             final EditText input = new EditText(this);
             alert.setView(input);
-
+            input.setText(card.getName());
+            input.setGravity(Gravity.CENTER);
+            input.setSelection(card.getName().length());//커서를 문자끝에 위치 시킴
             alert.setPositiveButton(R.string.card_edit_text_dialog_ok_button_text, new DialogInterface.OnClickListener() {
                 public void onClick(DialogInterface dialog, int whichButton) {
                     String value = input.getText().toString();
@@ -247,7 +270,13 @@ public class CardEditActivity extends AppCompatActivity {
         // return data
         if (intentResultCode == RESULT_OK) {
             Intent data = new Intent();
-            data.putExtra(IntentExtrasName.RETURN_DATA, card);
+            switch (intentRequestCode) {
+                case IntentRequestCode.CARD_ADD:
+                case IntentRequestCode.CARD_EDIT:
+                case IntentRequestCode.CARD_DELETE:
+                    data.putExtra(IntentExtrasName.RETURN_DATA, card);
+            }
+
             setResult(intentResultCode, data);
         } else {
             setResult(intentResultCode);
